requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pymongo==4.12.0
pydantic-settings==2.8.1
httpx==0.28.1
pytest==8.3.5
telethon==1.42.0
imageio-ffmpeg==0.6.0
opencv-python==4.13.0.92
rapidocr-onnxruntime==1.2.3
yt-dlp==2026.3.17
youtube-transcript-api==1.2.4
tzdata
